room_rate_management/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
